@@ -1,6 +1,6 @@
 {
   "name": "direflow-component",
-  "version": "1.0.11",
+  "version": "1.0.13",
   "description": "Create Web Components using React",
   "main": "dist/index.js",
   "author": "Silind Software",
@@ -19,21 +19,27 @@
   },
   "dependencies": {
     "@svgr/webpack": "4.3.2",
+    "@types/react": "16.9.3",
     "@types/webfontloader": "^1.6.29",
     "@webcomponents/webcomponentsjs": "^2.3.0",
+    "babel-plugin-styled-components": "^1.10.6",
     "event-hooks-webpack-plugin": "2.1.4",
     "html-webpack-externals-plugin": "^3.8.0",
     "react-shadow": "17.1.3",
-    "webfontloader": "^1.6.28"
+    "to-string-loader": "1.1.5",
+    "webfontloader": "^1.6.28",
+    "webpack-filter-warnings-plugin": "^1.2.1"
   },
   "devDependencies": {
-    "typescript": "3.6.2"
+    "typescript": "3.7.2"
   },
   "peerDependencies": {
     "@types/react": "16.9.3",
     "@types/react-dom": "16.9.1",
+    "@types/styled-components": "^4.1.22",
     "react": "16.10.1",
     "react-dom": "16.10.1",
-    "style-it": "2.1.4"
+    "style-it": "2.1.4",
+    "styled-components": "^4.4.1"
   }
 }
