@@ -1,6 +1,6 @@
 {
   "name": "direflow-project",
-  "version": "1.0.11",
+  "version": "1.0.13",
   "description": "Create Web Components using React",
   "main": "dist/index.js",
   "author": "Silind Software",
@@ -29,6 +29,6 @@
     "webpack-merge": "4.2.1"
   },
   "devDependencies": {
-    "typescript": "3.6.2"
+    "typescript": "3.7.2"
   }
 }
