@@ -1,18 +1,18 @@
 module.exports = {
   direflowMetadata: {
-    title: "%name-snake%",
-    description: "%description%",
-    type: "%setup-type%",
-    createVersion: "%install-version%"
+    title: '%name-snake%',
+    description: '%description%',
+    type: '%setup-type%',
+    createVersion: '%install-version%',
   },
   plugins: [
     {
       name: 'font-loader',
       options: {
         google: {
-          families: ['PT Sans', 'Noto Sans JP']
-        }
-      }
-    }
-  ]
-}
+          families: ['PT Sans', 'Noto Sans JP'],
+        },
+      },
+    },
+  ],
+};
